[fix] modal menu вынесено в отдельный компонент
