+++ src/scripts/components/modal/modal.hbs
@@ -1,0 +1,8 @@
+<link rel="stylesheet" href="/src/scripts/components/modal/modal.css">
+
+<div id="{{id}}" class="modal__main" tabindex="-1">
+    <div class="modal__container">
+        <span class="btn__close" data-action="close">&times;</span>
+
+    </div>
+</div>
